schleuder: switch default keyserver to keys.openpgp.org
Schleuder is able to fetch and update keys from a keyserver (pool).

Traditionally, Schleuder relied on the SKS keyservers. However, due to
various reasons, this network is broken, unreliable, and, depending on
the perspective, dying (slowly).

This commit changes the default keyserver to keys.openpgp.org, which is,
in contrast to SKS, a single server. It is validating, people need to
opt-in to have their mail address published along their key. GPG in
Debian ships a patch to make it work with keys without uid.

This change aligns with a common switch within the GPG ecosystem:
Debian, Mailvelope and Enigmail, just to name a few, switched their
default as well.

This commit might break x-fetch-key, as GPG refuses to import new keys
without uids, in contrast to updates of existing keys, as described
above.